map and random files fixed

+++ Random.cpp
@@ -2,12 +2,12 @@
 #include "Random.h"
 #include <cstdlib>
 // Returns a random number in r.
-int Random(Range r)
+float Random(Range r)
 {
 	return r.mLow + rand() % ((r.mHigh + 1) - r.mLow);
 }
 // Returns a random number in [low, high].
-int Random(int low, int high)
+float Random(float low, float high)
 {
 	return low + rand() % ((high + 1) - low);
 }
