+++ Random.h
@@ -2,6 +2,6 @@
 #ifndef RANDOM_H
 #define RANDOM_H
 #include "Range.h"
-int Random(Range r);
-int Random(int a, int b);
+float Random(Range r);
+float Random(float a, float b);
 #endif // RANDOM_H
